@@ -10,12 +10,12 @@ plugins {
 
 android {
     namespace 'com.example.scanqrapp'
-    compileSdk 32
+    compileSdk 33
 
     defaultConfig {
         applicationId "com.example.scanqrapp"
         minSdk 23
-        targetSdk 32
+        targetSdk 33
         versionCode 1
         versionName "1.0"
 
@@ -43,15 +43,15 @@ android {
 
 dependencies {
 
-    implementation 'androidx.appcompat:appcompat:1.5.1'
-    implementation 'com.google.android.material:material:1.7.0'
+    implementation 'androidx.appcompat:appcompat:1.6.1'
+    implementation 'com.google.android.material:material:1.8.0'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
     implementation 'androidx.lifecycle:lifecycle-livedata-ktx:2.5.1'
     implementation 'androidx.lifecycle:lifecycle-viewmodel-ktx:2.5.1'
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
 
     //firebase
-    implementation 'com.google.firebase:firebase-abt:21.1.0'
+    implementation 'com.google.firebase:firebase-abt:21.1.1'
     implementation 'com.google.firebase:firebase-database:20.1.0'
     implementation 'com.google.firebase:firebase-storage:20.1.0'
     implementation 'com.firebase:firebase-client-android:2.5.2'
@@ -60,21 +60,22 @@ dependencies {
     implementation fileTree(dir: 'D:\\AppGallery\\appScan-\\app\\libs', include: ['*.aar', '*.jar'], exclude: [])
     implementation files('libs/poi-3.7.jar')
 
-    implementation 'androidx.core:core:1.8.0'
-    implementation 'androidx.annotation:annotation:1.5.0'
+    implementation 'androidx.core:core:1.9.0'
+    //implementation 'androidx.annotation:annotation:1.6.0'
     implementation 'com.google.firebase:firebase-auth:21.1.0'
 
     implementation 'com.google.firebase:firebase-crashlytics:18.3.5'
+   // implementation 'androidx.annotation:annotation-jvm:1.6.0'
 
 
-    testImplementation 'junit:junit:4.13.2'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.4'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.0'
+    testImplementation 'org.junit.jupiter:junit-jupiter:5.9.2'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.5'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
 
     // Navigation
     implementation 'androidx.navigation:navigation-fragment-ktx:2.5.3'
     implementation 'androidx.navigation:navigation-ui-ktx:2.5.3'
-    implementation "androidx.fragment:fragment-ktx:1.5.4"
+    implementation "androidx.fragment:fragment-ktx:1.5.5"
 
     implementation 'com.github.yuriy-budiyev:code-scanner:2.3.2'
     implementation 'me.dm7.barcodescanner:zxing:1.9.13'
